# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Auth Service API
  version: "1.0"
  description: REST API для аутентификации и управления токенами
servers:
  - url: http://localhost:8080
paths:
  /auth/token:
    get:
      summary: Получить токены
      tags:
        - Auth
      parameters:
        - in: query
          name: user_id
          required: true
          schema:
            type: string
          description: UUID пользователя
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
        '400':
          description: Неверный запрос
        '500':
          description: Ошибка сервера

  /auth/refresh:
    post:
      summary: Обновить токены
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                refresh_token:
                  type: string
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
        '400':
          description: Неверный запрос
        '401':
          description: Неавторизовано
        '500':
          description: Ошибка сервера

  /auth/user:
    get:
      summary: Получить текущий user_id
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
        '401':
          description: Неавторизовано

  /auth/logout:
    post:
      summary: Деавторизация
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Успешно деавторизован
        '401':
          description: Неавторизовано

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT